Update OpenSSL to the new stable branch and remove CF patch
The cloudflare patch allows non AES-NI devices to use ChaCha20 instead of AES. ARMv8 devices are optimized for AES instructions, so there is no visible speed difference.
As I'm removing the patch, it allows me to update to OpenSSL 1.1.0 branch. This version of OpenSSL have ChaCha built-in (but like any other cipher).